Fix compilation error.
I didn't add this file in my last commit -- javac complains about an
implicit narrowing conversion from long -> int.

+++ src/net/sourceforge/kolmafia/request/CharSheetRequest.java
@@ -378,7 +378,7 @@ public class CharSheetRequest extends GenericRequest {
     }
 
     stats[0] = StringUtilities.parseInt(tokens[++pos]);
-    int base = CharSheetRequest.retrieveBase(tokens[++pos], (int) stats[0]);
+    int base = (int) CharSheetRequest.retrieveBase(tokens[++pos], (int) stats[0]);
 
     int subPoints = 0;
